moved all customer stylesheets into new styles directory
