+++ src/styles/customers/detail.css
@@ -1,0 +1,7 @@
+.customer-details--container {
+    border: 3px solid black;
+    border-radius: 8px;
+    max-width: 270px;
+    padding-left: 20px;
+    margin: 50px auto;
+}
