+++ src/styles/customers/list.css
@@ -1,0 +1,14 @@
+.customersContainer {
+    padding: 10px;
+    min-width: 270px;
+    max-width: 270px;
+    max-height: 250px;
+    overflow: scroll;
+    -ms-overflow-style: none;  /* IE and Edge */
+    scrollbar-width: none;  /* Firefox */
+}
+
+/* Hide scrollbar for Chrome/Safari/Opera */
+.customersContainer::-webkit-scrollbar {
+    display: none;
+}
